Separating code from project-settings into subcomponents
