+++ src/app/base/entities/project.ts
@@ -7,4 +7,6 @@ export interface Project {
   projectType: string;
   description?: string;
   numDataScaleUploaded?: number;
+  tokenizer?: string;
+  containsUniqueAttribute?: boolean;
 }
